requirements.txt
langchain>=0.0.206
faiss-cpu==1.7.4
GitPython==3.1.31
unstructured>=0.6.7
sentence-transformers==2.2.2
sentencepiece==0.1.98
rwkv==0.7.5
tokenizer==3.4.2
redis==4.5.5
tornado>=6.3.2
streamlit>=1.24.0
jq==1.4.1
